Split out WebBot superclass
